Use Lua API to speed up note browsing
The initial idea was to rely on `grepprg` to generate the note browser
content, however, the implementation was poor and the use of external
program calls were adding extra over head to note fetch (especially on
Windows.) I was also unable to properly support `ripgrep` and `ugrep`
because of issue #9.

Note browsing must be as fast as possible so I don't waste time waiting
for the browser, it's been annoying but hadn't had the time to address
the issue.

Now with this commit, I'm relying on the Lua platform calls and Vim
functions to get the note contents. I'm also doing an in-memory caching
for the notes which will be refreshed the next time the notes are
fetched if the note was modified.

Hopefully this will be much better.

Closes issues #2, #9, #10, and #12.