Allow tags if they are not part of a sentence
It used to be that a URL like `https://link.to/content#anchor` in a note
would result in a tag (`#anchor`) listed in the browser for that note.
That was confusing as it's not meant to be an anchors. Hopefully, this
commit fixes that problem. I don't know if I can change the pattern for
this to fix it instead of looking up the previous character. My pattern
skills are not that great for Lua.