1. id scheme: Adds id_format option.
The id_format option can be either a string or a function.
If it is a string it is treated as a format string for strftime, as is the
current use. (In this case exposing this option is like exposing the
NOTE_ID_STRFTIME_FORMAT.)
If it is a function it is called with the provide parent_id (possibly
nil) and the list of ids. It should return a new id for the new file.