implemented extensions (#25)
* 1. id scheme: Adds id_format option.

The id_format option can be either a string or a function.
If it is a string it is treated as a format string for strftime, as is the
current use. (In this case exposing this option is like exposing the
NOTE_ID_STRFTIME_FORMAT.)
If it is a function it is called with the provide parent_id (possibly
nil) and the list of ids. It should return a new id for the new file.

* 2. get_notes(): documents the get_notes functions.

* 3. get_tags(): documents the get_tags()

* Bonus: adds "contains" function.

* Suggested spelling correction.

Co-authored-by: Furkan Üzümcü <furkanuzumcu@gmail.com>

* Merge of Furkanzmc into branch: spelling and better commands.

* Ran stylua

* Removes ids argument to id_format. Updates doc.

---------

Co-authored-by: Furkan Üzümcü <furkanuzumcu@gmail.com>